requirements.txt
aiohappyeyeballs==2.4.3 ; python_version >= "3.11" and python_version < "3.13"
aiohttp==3.11.7 ; python_version >= "3.11" and python_version < "3.13"
aiosignal==1.3.1 ; python_version >= "3.11" and python_version < "3.13"
astunparse==1.6.3 ; python_version >= "3.11" and python_version < "3.13"
attrs==24.2.0 ; python_version >= "3.11" and python_version < "3.13"
cachetools==5.5.0 ; python_version >= "3.11" and python_version < "3.13"
certifi==2024.8.30 ; python_version >= "3.11" and python_version < "3.13"
charset-normalizer==3.4.0 ; python_version >= "3.11" and python_version < "3.13"
click==8.1.7 ; python_version >= "3.11" and python_version < "3.13"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "3.13" and (sys_platform == "win32" or platform_system == "Windows")
db-dtypes==1.3.1 ; python_version >= "3.11" and python_version < "3.13"
decorator==5.1.1 ; python_version >= "3.11" and python_version < "3.13"
dlt==1.4.0 ; python_version >= "3.11" and python_version < "3.13"
duckdb==1.1.3 ; python_version >= "3.11" and python_version < "3.13"
frozenlist==1.5.0 ; python_version >= "3.11" and python_version < "3.13"
fsspec==2024.10.0 ; python_version >= "3.11" and python_version < "3.13"
gcsfs==2024.10.0 ; python_version >= "3.11" and python_version < "3.13"
gitdb==4.0.11 ; python_version >= "3.11" and python_version < "3.13"
gitpython==3.1.43 ; python_version >= "3.11" and python_version < "3.13"
giturlparse==0.12.0 ; python_version >= "3.11" and python_version < "3.13"
google-api-core==2.23.0 ; python_version >= "3.11" and python_version < "3.13"
google-api-core[grpc]==2.23.0 ; python_version >= "3.11" and python_version < "3.13"
google-auth-oauthlib==1.2.1 ; python_version >= "3.11" and python_version < "3.13"
google-auth==2.36.0 ; python_version >= "3.11" and python_version < "3.13"
google-cloud-bigquery==3.27.0 ; python_version >= "3.11" and python_version < "3.13"
google-cloud-core==2.4.1 ; python_version >= "3.11" and python_version < "3.13"
google-cloud-storage==2.18.2 ; python_version >= "3.11" and python_version < "3.13"
google-crc32c==1.6.0 ; python_version >= "3.11" and python_version < "3.13"
google-resumable-media==2.7.2 ; python_version >= "3.11" and python_version < "3.13"
googleapis-common-protos==1.66.0 ; python_version >= "3.11" and python_version < "3.13"
grpcio-status==1.68.0 ; python_version >= "3.11" and python_version < "3.13"
grpcio==1.68.0 ; python_version >= "3.11" and python_version < "3.13"
hexbytes==1.2.1 ; python_version >= "3.11" and python_version < "3.13"
humanize==4.11.0 ; python_version >= "3.11" and python_version < "3.13"
idna==3.10 ; python_version >= "3.11" and python_version < "3.13"
jsonpath-ng==1.7.0 ; python_version >= "3.11" and python_version < "3.13"
loguru==0.7.2 ; python_version >= "3.11" and python_version < "3.13"
makefun==1.15.6 ; python_version >= "3.11" and python_version < "3.13"
multidict==6.1.0 ; python_version >= "3.11" and python_version < "3.13"
numpy==2.1.3 ; python_version >= "3.11" and python_version < "3.13"
oauthlib==3.2.2 ; python_version >= "3.11" and python_version < "3.13"
orjson==3.10.11 ; python_version >= "3.11" and python_version < "3.13" and platform_python_implementation != "PyPy"
packaging==24.2 ; python_version >= "3.11" and python_version < "3.13"
pandas==2.2.3 ; python_version >= "3.11" and python_version < "3.13"
pathvalidate==3.2.1 ; python_version >= "3.11" and python_version < "3.13"
pendulum==3.0.0 ; python_version >= "3.11" and python_version < "3.13"
pluggy==1.5.0 ; python_version >= "3.11" and python_version < "3.13"
ply==3.11 ; python_version >= "3.11" and python_version < "3.13"
propcache==0.2.0 ; python_version >= "3.11" and python_version < "3.13"
proto-plus==1.25.0 ; python_version >= "3.11" and python_version < "3.13"
protobuf==5.28.3 ; python_version >= "3.11" and python_version < "3.13"
pyarrow==18.0.0 ; python_version >= "3.11" and python_version < "3.13"
pyasn1-modules==0.4.1 ; python_version >= "3.11" and python_version < "3.13"
pyasn1==0.6.1 ; python_version >= "3.11" and python_version < "3.13"
python-dateutil==2.9.0.post0 ; python_version >= "3.11" and python_version < "3.13"
python-dotenv==1.0.1 ; python_version >= "3.11" and python_version < "3.13"
pytz==2024.2 ; python_version >= "3.11" and python_version < "3.13"
pyyaml==6.0.2 ; python_version >= "3.11" and python_version < "3.13"
requests-oauthlib==2.0.0 ; python_version >= "3.11" and python_version < "3.13"
requests==2.32.3 ; python_version >= "3.11" and python_version < "3.13"
requirements-parser==0.11.0 ; python_version >= "3.11" and python_version < "3.13"
rsa==4.9 ; python_version >= "3.11" and python_version < "3.13"
semver==3.0.2 ; python_version >= "3.11" and python_version < "3.13"
setuptools==75.6.0 ; python_version >= "3.11" and python_version < "3.13"
simplejson==3.19.3 ; python_version >= "3.11" and python_version < "3.13"
six==1.16.0 ; python_version >= "3.11" and python_version < "3.13"
smmap==5.0.1 ; python_version >= "3.11" and python_version < "3.13"
sqlglot==25.32.0 ; python_version >= "3.11" and python_version < "3.13"
tenacity==9.0.0 ; python_version >= "3.11" and python_version < "3.13"
tomlkit==0.13.2 ; python_version >= "3.11" and python_version < "3.13"
tower-cli==0.1.19 ; python_version >= "3.11" and python_version < "3.13"
tqdm==4.67.0 ; python_version >= "3.11" and python_version < "3.13"
types-setuptools==75.5.0.20241122 ; python_version >= "3.11" and python_version < "3.13"
typing-extensions==4.12.2 ; python_version >= "3.11" and python_version < "3.13"
tzdata==2024.2 ; python_version >= "3.11" and python_version < "3.13"
urllib3==2.2.3 ; python_version >= "3.11" and python_version < "3.13"
wheel==0.45.1 ; python_version >= "3.11" and python_version < "3.13"
win-precise-time==1.4.2 ; python_version >= "3.11" and python_version < "3.13" and os_name == "nt"
win32-setctime==1.1.0 ; python_version >= "3.11" and python_version < "3.13" and sys_platform == "win32"
yarl==1.18.0 ; python_version >= "3.11" and python_version < "3.13"


#
# Local dependencies added manually
#
./.dist/dlt_plus-0.2.6.tar.gz
